Rename BTHome MiThermometer component and add voltage icon
